MAG-12: make bottom line of level to deal damage

+++ Assets/Scripts/Health.cs
@@ -6,9 +6,9 @@ public class Health : NetworkBehaviour {
     [SerializeField] private int maxHealth = 100;
 
     [SyncVar(hook = nameof(HandleHealthUpdated))]
-    private int currentHealth;
+    private float currentHealth;
 
-    public event Action<int, int> ClientOnHealthUpdated;
+    public event Action<float, float> ClientOnHealthUpdated;
 
     #region Server
 
@@ -17,7 +17,7 @@ public class Health : NetworkBehaviour {
     }
 
     [Server]
-    public void DealDamage(int damageAmount) {
+    public void DealDamage(float damageAmount) {
         if (currentHealth == 0) { return; }
 
         currentHealth = Mathf.Max(currentHealth - damageAmount, 0);
@@ -31,7 +31,7 @@ public class Health : NetworkBehaviour {
 
     #region Client
 
-    private void HandleHealthUpdated(int oldValue, int newValue) {
+    private void HandleHealthUpdated(float oldValue, float newValue) {
         ClientOnHealthUpdated?.Invoke(newValue, maxHealth);
     }
